add appointment messages && modify some usage of functions

+++ Pages/HomeForm.cs
@@ -1,6 +1,7 @@
 ﻿using System.ComponentModel;
-using System.Linq;
+using System.Globalization;
 using CompanyScheduler.Models;
+using CompanyScheduler.Pages.Calendar;
 using CompanyScheduler.Pages.Customers;
 using Microsoft.EntityFrameworkCore;
 
@@ -8,17 +9,13 @@ namespace CompanyScheduler.Pages;
 
 /// <summary>
 /// Form handling main activities
-///
-/// TODO:
-///     Add Exception Handling add update delete
-///     Add report functions
 /// </summary>
-
-
 public partial class HomeForm : Form
 {
+    private readonly DateTime currentTime = DateTime.UtcNow;
+    private readonly GregorianCalendar calendar = new();
     private BindingList<Appointment>? appointments;
-    private Appointment? _selectedAppointment = null;
+    private Appointment? _selectedAppointment;
 
     public User User { get; private set; }
 
@@ -26,6 +23,7 @@ public partial class HomeForm : Form
     {
         InitializeComponent();
         LoadAppointments();
+        CheckSoonAppointments(15);
         User = user;
     }
 
@@ -33,6 +31,43 @@ public partial class HomeForm : Form
     {
         using var context = new ClientScheduleContext();
         appointments = [.. context.Appointments.Include(a => a.User)];
+    }
+
+    private void CheckSoonAppointments(int span)
+    {
+        using var context = new ClientScheduleContext();
+
+        foreach (var appointment in context.Appointments)
+        {
+            var minutesLeft = Math.Abs((appointment.Start - currentTime).TotalMinutes);
+
+            if (minutesLeft <= span)
+            {
+                string message = $"Appointment {appointment.Title} starts in {minutesLeft}";
+                string caption = $"Appointment Soon!";
+                MessageBoxButtons buttons = MessageBoxButtons.OK;
+                MessageBox.Show(message, caption, buttons);
+            }
+        }
+    }
+
+    private void AppointmentListView_IndexChanged(object sender, EventArgs e)
+    {
+        if (appointmentListView.SelectedItems.Count > 0)
+        {
+            _selectedAppointment = appointmentListView.SelectedItems[0].Tag as Appointment;
+        }
+    }
+
+    private void CalendarButton_Clicked(object sender, EventArgs e)
+    {
+        if (_selectedAppointment is null)
+            return;
+
+        var CalendarForm = new CalendarForm(this, calendar, User);
+
+        CalendarForm.Show();
+        Hide();
     }
 
     private void CreateCustomerButton_Clicked(object sender, EventArgs e)
@@ -58,16 +93,18 @@ public partial class HomeForm : Form
         if (_selectedAppointment is null || _selectedAppointment.Customer is null)
             return;
 
-        var UpdateCustomer = new CustomerUpdateForm(User, _selectedAppointment.Customer, this);
+        var tbdAppointment = _selectedAppointment;
 
-        using (var context = new ClientScheduleContext())
+        var UpdateCustomer = new CustomerUpdateForm(User, _selectedAppointment.Customer, this);
+        UpdateCustomer.CustomerUpdated += (sender, customer) =>
         {
+            using var context = new ClientScheduleContext();
             context.Appointments.Remove(_selectedAppointment);
-            UpdateCustomer.CustomerUpdated += (sender, customer) =>
-                _selectedAppointment.Customer = customer;
+
+            _selectedAppointment.Customer = customer;
             context.Appointments.Add(_selectedAppointment);
             context.SaveChanges();
-        }
+        };
 
         UpdateCustomer.Show();
         Hide();
@@ -75,30 +112,42 @@ public partial class HomeForm : Form
 
     private void DeleteCustomerButton_Clicked(object sender, EventArgs e)
     {
-        if (_selectedAppointment is null || _selectedAppointment.User is null)
-            return;
-
         using var context = new ClientScheduleContext();
+        if (
+            _selectedAppointment is null
+            || _selectedAppointment.User is null
+            || !context.Appointments.Contains(_selectedAppointment)
+        )
+        {
+            string message = "Failed to Delete Customer";
+            string caption = "Missing Customer";
+            MessageBoxButtons messageBoxButtons = MessageBoxButtons.OK;
+            MessageBox.Show(message, caption, messageBoxButtons);
+            return;
+        }
+
         context.Appointments.Remove(_selectedAppointment);
         context.SaveChanges();
     }
 
     private void AppointmentsButton_Clicked(object sender, EventArgs e) { }
 
-    private int GenerateAppointmentTypesByMonth(DateOnly date)
-    {
-        throw new NotImplementedException();
-    }
+    // Report generating functions
+    private IQueryable<string>? GenerateAppointmentTypesByMonth(DateOnly date) =>
+        new ClientScheduleContext()
+            .Appointments.Where(appointment => appointment.Start.Month == date.Month)
+            .Select(appointment => appointment.Type)
+            .Distinct();
 
-    private void GenerateSchedule()
-    {
-        throw new NotImplementedException();
-    }
+    private IQueryable<List<Appointment>>? GenerateSchedule() =>
+        new ClientScheduleContext()
+            .Appointments.GroupBy(appointment => appointment.User)
+            .Select(appointments => appointments.ToList());
 
-    private void GeneratePerCity(City city)
-    {
-        throw new NotImplementedException();
-    }
+    private IQueryable<bool>? GenerateAllCustomerWithAppointments() =>
+        new ClientScheduleContext()
+            .Customers.Select(customer => customer.Appointments.Count != 0)
+            .Distinct();
 
     private void QuitButton_Clicked(object sender, EventArgs e) => Environment.Exit(0);
 }
